requirements.txt
langchain==0.3.27
langchain-community==0.3.31
langchain-huggingface==0.3.1
chromadb==1.1.1
streamlit==1.50.0
py==1.11.1
python-dotenv==1.1.1
pypdf==6.1.1
sentence-transformers==5.1.1
transformers==4.57.1